Renamed the base tests
