+++ tests/unit/data/base/__init__.py
@@ -3,10 +3,10 @@
 # Copyright (C) 2023 Intel Corporation
 # SPDX-License-Identifier: Apache-2.0
 
-from .test_datamodule import _TestAnomalibDataModule
-from .test_depth import _TestAnomalibDepthDatamodule
-from .test_image import _TestAnomalibImageDatamodule
-from .test_video import _TestAnomalibVideoDatamodule
+from .datamodule import _TestAnomalibDataModule
+from .depth import _TestAnomalibDepthDatamodule
+from .image import _TestAnomalibImageDatamodule
+from .video import _TestAnomalibVideoDatamodule
 
 __all__ = [
     "_TestAnomalibDataModule",
